rubocop

+++ app/controllers/sessions_controller.rb
@@ -4,15 +4,13 @@ class SessionsController < ApplicationController
       @user = User.from_omniauth(request.env['omniauth.auth'])
       session[:user_id] = @user.uid
     rescue
-      flash[:warning] = "You are not a Klarna employee. Please see www.klarnaisrael.com for more details"
+      flash[:warning] = 'You are not a Klarna employee. Please see www.klarnaisrael.com for more details'
     end
     redirect_to root_path
   end
 
   def destroy
-    if current_user
-      session.delete(:user_id)
-    end
+    session.delete(:user_id) if current_user
     redirect_to root_path
   end
 end
